requirements.txt
Flask==1.1.2
google-cloud-storage==1.35.1
google-cloud-logging==2.2.0
google-cloud-error-reporting==1.1.0
gunicorn==20.0.4
six==1.15.0
cassandra-driver==3.24.0